Better error messages
This commit improves the error messages that are emitted if something
goes wrong by including some details like field count and field names into the panic
message. This requires changing the rust edition for this crate to 2018
(or older) to get back the old sematics for panic! which allowed to pass
a string as first argument. This string then needs to be manually
constructed as most relevant functions (like `to_string()` and
`Display`, etc) are not callable in const contexts yet. This also
requires some unsafe const code to construct these strings. That should
be fine as the compiler checks if that code is "fine" by executing it
through miri. So even if some casts there look broken it seems to work.

This commit also tries to use `#[track_caller]` but that doesn't seem to
work as expected yet.